requirements.txt
Django==2.1
pytz==2018.9
requests==2.24.0
newsapi-python